+++ lib/wasix/src/journaling/effector/syscalls/sock_send_to.rs
@@ -41,7 +41,7 @@ impl JournalEffector {
                         data: Cow::Borrowed(buf.as_ref()),
                         addr,
                         is_64bit: M::is_64bit(),
-                        flags: si_flags as u16,
+                        flags: si_flags,
                     })
                     .await
                     .map_err(map_snapshot_err)?;
